chore: Refactored to separate client and provider
Also defined enums for name transformers and formats.
