+++ config/crds/bases/external-secrets.io_secretstores.yaml
@@ -2032,10 +2032,22 @@ spec:
                       format:
                         description: Format enables the downloading of secrets as
                           a file (string)
+                        enum:
+                        - json
+                        - dotnet-json
+                        - env
+                        - yaml
+                        - docker
                         type: string
                       nameTransformer:
                         description: Environment variable compatible name transforms
                           that change secret names to a different format
+                        enum:
+                        - upper-camel
+                        - camel
+                        - lower-snake
+                        - tf-var
+                        - dotnet-env
                         type: string
                       project:
                         description: Doppler project (required if not using a Service
